pcap: fix TINYINT UNSIGNED decode; add APP_TZ_OFFSET_HOURS
pcap:

* cdr_tar_part.type is TINYINT UNSIGNED — decode as u8, not i8

   (sqlx refuses i8 for unsigned TINYINT)

tz:

* new APP_TZ_OFFSET_HOURS env var (integer hours from UTC)

* Config::tz() returns a chrono::FixedOffset

* default today filter window and day_label respect the tz

* distinct-values 7-day lookback also uses tz

* 2 unit tests covering tz window behaviour across offsets

+++ src/routes/pcap.rs
@@ -55,7 +55,8 @@ impl TarPartRow {
         Ok(Self {
             calldate: r.try_get("calldate")?,
             pos: r.try_get::<u64, _>("pos")?,
-            type_: r.try_get::<i8, _>("type")? as u8,
+            // `type` is TINYINT UNSIGNED in voipmonitor's schema.
+            type_: r.try_get::<u8, _>("type")?,
         })
     }
 }
